Style : 1주차 과제 컴포넌트들에 대한 디렉토리 구조 변화
아토믹 디자인 패턴을 사용하여 제작했던 컴포넌트들을 분리해줬다.
