+++ my-app/src/atoms/CarouselItem.jsx
