+++ my-app/src/molecules/Toast.jsx
@@ -1,44 +1,6 @@
 import { useCallback, useEffect, useState } from 'react'
 import styles from '../styles/Toast.css'
-
-const ToastBox = ({ toastlist, setList }) => {
-  const deleteToast = useCallback(id => {
-    const toastListItem = toastlist.filter(e => e.id !== id)
-    setList(toastListItem)
-  }, [toastlist, setList])
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      if (toastlist.length) {
-        deleteToast(toastlist[0].id)
-      }
-    }, 3000);
-
-    return () => {
-      clearInterval(interval)
-    }
-  }, [toastlist, deleteToast])
-
-  return (
-    <div className='container'>
-      {
-        toastlist.map((toast, i) => (
-          <div 
-            key={i} 
-            style={{ backgroundColor: toast.backgroundColor }}
-            className='notification'
-          >
-            <button className='button' onClick={() => deleteToast(toast.id)}>X</button>
-            <div>
-              <p className='title'>{toast.title}</p>
-              <p className='description'>{toast.description}</p>
-            </div>
-          </div>
-        ))
-      }
-    </div>
-  )
-}
+import ToastBox from '../atoms/ToastBox'
 
 const Toast = ({isValid, task}) => {
   const [list, setList] = useState([]);
